Moved Home/ProjectSubmission into routes, removed pages dir.
